+++ src/index.ts
@@ -1,67 +1,24 @@
-import fetch from "node-fetch";
-import get from "lodash.get";
-import parseData from "./parseInitialData";
+import { buildPlaylist } from "./build-playlist";
+import { getPlaylistHtml } from "./get-playlist-html";
+import parseData from "./parse-initial-data";
+import { Playlist } from "./types";
+export * from "./types";
 
-export type Video = {
-  id: string;
-  title: string;
-  name: string;
-};
-export type Playlist = {
-  title: string;
-  playlist: Video[];
-};
-
-const getPlaylistUrl = (id: string): string =>
-  `https://www.youtube.com/playlist?list=${id}`;
-
-// No way I'm typing the ytInitialData
-export const extractItems = (data: object): Video[] => {
-  const items = get(
-    data,
-    "contents.twoColumnBrowseResultsRenderer.tabs[0].tabRenderer.content.sectionListRenderer.contents[0].itemSectionRenderer.contents[0].playlistVideoListRenderer.contents"
-  );
-
-  return items.reduce((acc: Video[], item: any) => {
-    if (typeof item.playlistVideoRenderer === "undefined") return acc;
-    const body = item.playlistVideoRenderer;
-    const id = body.videoId;
-    return [
-      ...acc,
-      {
-        id,
-        name: get(body, "title.runs[0].text") || "",
-        url: `https://youtube.com/watch?v=${id}`,
-      },
-    ];
-  }, []);
-};
-
-export const createPayload = (data: object): Playlist => {
-  const maybeTitle = get(data, "metadata.playlistMetadataRenderer.title") || "";
-  let payload: Playlist | null = null;
-
-  payload = {
-    title: maybeTitle.trim(),
-    playlist: extractItems(data) || [],
-  };
-  return payload;
-};
+/**
+ * Try and scrape the playlist for a given ID
+ *
+ * @param playlistId - ID of the playlist we wish to scrape
+ */
 export default async function scrapePlaylist(
   playlistId: string
 ): Promise<Playlist> {
-  const url = getPlaylistUrl(playlistId);
-  const response = await fetch(url);
-  const html = await response.text();
-  if (!response.ok) {
-    throw new Error(`Unable to fetch playlist with ID: ${playlistId}`);
-  }
+  const html = await getPlaylistHtml(playlistId);
   const data = parseData(html);
   if (data === null) {
     throw new Error("Unable to parse ytInitialData");
   }
   try {
-    return createPayload(data);
+    return buildPlaylist(data);
   } catch (error) {
     throw new Error("Unable to parse body");
   }
