+++ src/types.ts
@@ -1,0 +1,9 @@
+export type Video = {
+  id: string;
+  title: string;
+  name: string;
+};
+export type Playlist = {
+  title: string;
+  playlist: Video[];
+};
